requirements.txt
aiohttp==3.4.1
aiohttp-graphql==1.0.0
aniso8601==3.0.2
async-timeout==3.0.0
asyncpg==0.17.0
attrs==18.2.0
chardet==3.0.4
graphene==2.1.3
graphql-core==2.1
graphql-relay==0.4.5
graphql-server-core==1.1.1
idna==2.7
idna-ssl==1.1.0
multidict==4.3.1
promise==2.1
pytest-runner==4.2
Rx==1.6.1
six==1.11.0
typing==3.6.6
yarl==1.2.6
